requirements.txt
Flask
requests
litellm
python-dotenv  # If you're using a .env file